test(osmrBackend): services
Reflect changes to abstraction.
